+++ docs/source/changes.rst
@@ -41,7 +41,7 @@ Bug Fixes:
 New Features
 
 - Add support for Python 3.9
-- Drop support for python 3.4
+- Drop support for python 3.5
 
 0.4.2
 ------